move browser_utils under screen_parsing module
